requirements_nn.txt
# generated by:
# python3 -m venv env_nn
# source env_nn/bin/activate
# pip install --upgrade pip
# pip install scipy==1.8.0 numpy==1.22.3 numba==0.55.2 matplotlib==3.5.2 h5py==3.6.0 joblib==1.1.0 jax==0.3.14 jaxlib==0.3.14 dm-haiku==0.0.5 optax==0.1.0
# pip freeze >> requirements_nn.txt
absl-py==1.1.0
chex==0.1.3
cycler==0.11.0
dm-haiku==0.0.5
dm-tree==0.1.7
etils==0.6.0
flatbuffers==2.0
fonttools==4.34.4
h5py==3.6.0
importlib-resources==5.8.0
jax==0.3.14
jaxlib==0.3.14
jmp==0.0.2
joblib==1.1.0
kiwisolver==1.4.3
llvmlite==0.38.1
matplotlib==3.5.2
numba==0.55.2
numpy==1.22.3
opt-einsum==3.3.0
optax==0.1.0
packaging==21.3
Pillow==9.2.0
pkg_resources==0.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.8.0
six==1.16.0
tabulate==0.8.10
toolz==0.12.0
typing-extensions==3.10.0.2
zipp==3.8.0
